Editor: Fix up editor and demo page without iframes
Continues some of the previous fixes (moving lib files to src/lib for
create-react-app, etc) for the editor portions of perseus, and gets a
draft version of the editor page working again.

Getting the iframe working again looks tricky; so for now I replaced it
:/. Unfortunately its previous implementation was very tied to the
specific html needed to set up a perseus, so to recreate it in a way
that is publishable to npm I might have to heavily re-engineer it.